# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "VIP Tutors Task Management API",
        "description": "A comprehensive task management system with user authentication, task CRUD operations, reordering capabilities, and admin functionality.",
        "contact": {
            "name": "VIP Tutors Development Team",
            "email": "admin@viptutors.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local development server"
        }
    ],
    "paths": {
        "/admin/dashboard-stats": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get admin dashboard statistics",
                "description": "Retrieve comprehensive statistics for the admin dashboard",
                "operationId": "c5e50101291795eec0b8f5a61cebc6a1",
                "responses": {
                    "200": {
                        "description": "Dashboard statistics retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "total_users": {
                                            "type": "integer",
                                            "example": 150
                                        },
                                        "admin_users": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "regular_users": {
                                            "type": "integer",
                                            "example": 145
                                        },
                                        "total_tasks": {
                                            "type": "integer",
                                            "example": 1250
                                        },
                                        "completed_tasks": {
                                            "type": "integer",
                                            "example": 800
                                        },
                                        "pending_tasks": {
                                            "type": "integer",
                                            "example": 450
                                        },
                                        "high_priority_tasks": {
                                            "type": "integer",
                                            "example": 75
                                        },
                                        "tasks_created_today": {
                                            "type": "integer",
                                            "example": 25
                                        },
                                        "tasks_created_this_week": {
                                            "type": "integer",
                                            "example": 150
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Admin access required"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your bearer token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}